+++ trunk/source/Common/HashEngineInternal.h
@@ -108,10 +108,15 @@ namespace HashEngineInternal
 			xxh3_128(),
 			crc32c(0),
 			openSslSha256(),
+			openSslSha384(),
 			openSslSha512(),
 			openSslSha3_256(),
+			openSslSha3_384(),
 			openSslSha3_512(),
+			openSslBlake2b_160(),
+			openSslBlake2b_256(),
 			openSslBlake2b_512(),
+			openSslBlake2s_128(),
 			openSslBlake2s_256(),
 			openSslShake128_256(),
 			openSslShake256_512()
@@ -122,10 +127,15 @@ namespace HashEngineInternal
 		~FileHashContexts()
 		{
 			HashRuntime::CleanupOpenSslEvpHashContext(&openSslSha256);
+			HashRuntime::CleanupOpenSslEvpHashContext(&openSslSha384);
 			HashRuntime::CleanupOpenSslEvpHashContext(&openSslSha512);
 			HashRuntime::CleanupOpenSslEvpHashContext(&openSslSha3_256);
+			HashRuntime::CleanupOpenSslEvpHashContext(&openSslSha3_384);
 			HashRuntime::CleanupOpenSslEvpHashContext(&openSslSha3_512);
+			HashRuntime::CleanupOpenSslEvpHashContext(&openSslBlake2b_160);
+			HashRuntime::CleanupOpenSslEvpHashContext(&openSslBlake2b_256);
 			HashRuntime::CleanupOpenSslEvpHashContext(&openSslBlake2b_512);
+			HashRuntime::CleanupOpenSslEvpHashContext(&openSslBlake2s_128);
 			HashRuntime::CleanupOpenSslEvpHashContext(&openSslBlake2s_256);
 			HashRuntime::CleanupOpenSslEvpHashContext(&openSslShake128_256);
 			HashRuntime::CleanupOpenSslEvpHashContext(&openSslShake256_512);
@@ -145,10 +155,15 @@ namespace HashEngineInternal
 		XXH3_state_t xxh3_128;
 		uint32_t crc32c;
 		HashRuntime::OpenSslEvpHashContext openSslSha256;
+		HashRuntime::OpenSslEvpHashContext openSslSha384;
 		HashRuntime::OpenSslEvpHashContext openSslSha512;
 		HashRuntime::OpenSslEvpHashContext openSslSha3_256;
+		HashRuntime::OpenSslEvpHashContext openSslSha3_384;
 		HashRuntime::OpenSslEvpHashContext openSslSha3_512;
+		HashRuntime::OpenSslEvpHashContext openSslBlake2b_160;
+		HashRuntime::OpenSslEvpHashContext openSslBlake2b_256;
 		HashRuntime::OpenSslEvpHashContext openSslBlake2b_512;
+		HashRuntime::OpenSslEvpHashContext openSslBlake2s_128;
 		HashRuntime::OpenSslEvpHashContext openSslBlake2s_256;
 		HashRuntime::OpenSslEvpHashContext openSslShake128_256;
 		HashRuntime::OpenSslEvpHashContext openSslShake256_512;